Preserve attribute order
Since we are storing attributes in a hash we cannot rely on any ordering
so if we later write a cfg file to disk the attributes may be ordered
differently. Normally this should not make any difference but there
might be probe configuration file where ordering is crucial and it also
makes testing a lot harder because the generated configuration files are
unpredictable.

We now save the order in which we read the attributes from disk so we
can preserve the order if we late have to sync the modified
configuration back to disk