Add new type: agentil_template
A template defines the monitoring tasks that should apply to a group of
systems or one individual system. Inside the probe GUI you can define
multiple non-system templates and assign one or more template to each
system. Assigning templates to a system will trigger the creation of a
system template that is specific to one system and contains the actual
monitoring rules.

You can now create both types of templates with puppet.