Add new type: agentil_system
systems are the core component of the agentil configuration and each
system contains references to a landscape and to the user that can be
used to connect to the system. A system also includes references to
assigned templates and to the actual system template that contains
the actual monitoring rules.

Systems can now finally be managed through puppet with the
`agentil_system` resource type.