Do only register a configuration file on first access
We do not add the configuration file to our internal array as soon as
register_config is parsed and do a leazy adding on the first call to
config.

This is mainly done to ease testing because we can now stub the config
method pretty easily to work with custom configuration files during
testing.