Add helper script restart_probe.sh
Every configuration change of a probe needs a restart of the effected
probe. Unfortunately a probes cannot be started directly from the
commandline because they would not appear in the infrastructure manager
and may start with a wrong environment. A probe has to be started
through the controller probe.

The controller probe has an API to stop and start a specific probe but
this API only works with a valid adminsitrative nimsoft account and a
password - things you do not want to expose on every nimsoft robot.

The helper script now uses a different way: It will simply send a
SIGTERM to the specified probe and will wait for it to die. Once the
probe is down we wait for the controller to restart the probe.