Convert IDs into integers
An `agentil_template` resource with a specified `jobs` or `monitors`
property was always considered out of sync because the provider returns
the current value as an array of integers while the specfied should
value is always an array of strings.

Convert the supplied `jobs` and `monitors` array to all integers.