Ignore order of ip addresses in agentil_system
The order in which the `sapbasis_agentil` probe stores ip addresses is
unpredictable. As a result puppet may rewrite your configuration file
every time you made modifications the `sapbasis_agentil` GUI. Since the
order in which the ip addresses appear in the configuration does not
matter, puppet should not care either.