Do not write trailing spaces for empty attributes
If an attribute has no value assigned and you open the probe in the
probe GUI and than close the GUI again, you'll notice that the attribute
appears in the config file but has no value assigned to it, so the line
looks like `attribute =` instead of `attribute = `. This is now also
correctly handled in the nimsoft class.