Do not intern empty strings
The nimsoft configuration can contain settings with no value, e.g.

    suspended =

where we'd normally expect a value of `yes` or `no`. If we instruct the
provider to intern such a value we fail on ruby < 1.9 because an empty
string has no symbol representation.

Symbols as property values should only be used if there is a limited set
of allowed values that this property can have (e.g. `yes` and `no`). In
that case an empty value does not make sense and should be treated as if
the corresponding property was absent.

Result: If the manifest author does define a resource with a property
set to a specific value and this value is empty in the current
configuration, puppet will explicitly set it.