Refactor nimsoft_disk custom type
The nimsoft_disk type does now also do value validation on certain
attributes like `active` and `missing`. As a result these are symbolized
by the type and now also have to be symbolized in the get methods of the
provider.

Also remove the `warning_enable` and `critical_enable` properties. We
will now automatically enable a threshold when a numeric value is
defined and deactive a threshold, when `warning`/`critical` is set to
`absent`.