Remove agentil_instance type
The current `agentil_instance` type was badly modeled so remove the
type. The instances of a system should really be a parameter of the
`agentil_system` type, instead of a separate resource. But to accomplish
that, the `agentil` utility classes will need some refactoring.

So in the end ability to manage the monitored instances with puppet
will return, but to refactor the code to allow implementing it in a
sane way is easier, if we first remove the type completly.