Refactor agentil util classes
The different agentil classes shared a lot of implementation and there a
deep relationships between the types (e.g. a landscape has assigned
systems and each system has assigned templates etc) but there was not
central registry that had access to all objects.

Create a new class Puppet::Util::Agentil that holds a map of landscapes,
systems, users, and templates and is responsible for adding and removing
instances.