Remove the agentil job helper class
The agentil job class was formerly used to describe a customized job of
a specific template. This caused more overhead as necessary since every
customized job has completly different values. As a result a template
only stores a hash of customized jobs with the configuration element as
a value and no job object now.

This way a provider can query a specific job-overwrite and has to make
use of the returnes structure on its own.