Fix ruby 1.8.7 spec test incompatibility
Ruby 1.8.7 does not allow trailing commas in a list. Ruby 1.8.7 also
does not return entries of a hash in an deterministic order. As a result
the order of attributes of a newly created provider instance was random
on ruby 1.8.7 and caused spec failures.

Always create the attributes in alphabetical order of the matching
resource property.