Create configuration conversion script
The new sapbasis_agentil probe does no longer use the nimsoft configuration
format to store the probe configuration and has switched to json.

As a result we have to touch every util class and every provider to
acknowledge the new configuration format. But we also have to rewrite all
fixtures that we have created.

Create a simple script that will take a configuration file as input and
writes the new json format to stdout. This way we should be able to
convert all our fixtures to match the new format so we can acutally test
if all our providers comply with the new json configuration.