Change util classes to handle json config
The sapbasis_agenitl probe uses the json format for storing the
configuraiton. This means that we have to rewrite all util classes that
represent a landscape, system and template in the puppet world.

During the converstion agentil also renamed a few of their structure
- landscape is now called system
- system is now called connector
- user is now called user_profile

We currently use the old term to be backward compatible towards the
provider classes but this may change in the future