Update README for Agentil probe 4.00
Before version 4, the sapbasis_agentil probe used the nimsoft
configuration file format for its configuration. This has changed and
the configuration format is now JSON based. Since we already the agentil
util class and all other classes to use this new configuration format,
it is no longer possible to use the puppet types to handle configuration
files prior to version 4.00.

Update the README to document that dependency