Remove the `monitors` property
Since version 4.00 an agentil template does no longer have a `monitors`
property. This has propably been removed because it can be derived from
the configured `jobs` and hence can be generated on the fly.

Since it's no longer an attribute we can read or set, remove it as a
property from the `agentil_template` type