Add default values when adding a system
In general the providers do only care about the attributes they know
about. This means that the `agentil_*` types can probably not do
everything that you can do in the probe UI, but on the other hand if
things change we only have to be concerned if the change includes an
attribute that we can activly manage through the `agentil_*` types.

This assumes that the `sapbasis_agentil` will set every attribute to
default values if they do not appear in the configuration file. This is
not true: If we create a new system and ommit certain parameters (that
the provider currrently cannot manage) the probe fails to start.

To prevent the probe failures set all attributes for a new system even
if we cannot change them later.