Add new system parameter `ccms_mode`
The sapbasis_agentil largely depends on CCMS and has two ways to report
CCMS errors. If strict mode is enabled (`CCMS_STRICT_MODE=true`), you'll
get one message for each system and each error. This might produce a lot
of events if large parts of CCMS fail but on the other hand can already
see which CCMS error is present, without opening the probe UI. If you
disable the strict mode and inable the aggregated mode, you'll only get
one message for each *system* telling you how many errors are currently
present on each system.

This produces less alarms but you have to open the probe UI to find out
which alarms exactly are present.

Prior to this commit there was no way to define the behaviour for a new
`agentil_system`.

Add a new parameter `ccms_mode` that understands the values `strict` and
`aggregated` to configure the system appropiatly.