Add nfs property to nimsoft_disk
In general nimsoft does not monitor network devices, event when they are
active. This allows you to set the default value for `active` to `yes`
and still ignore network filesystems. This is important if a NFS share
is mounted on houndreds of servers because once this device exceeds a
filesystem threshold, you'd receive an alarm from every server that has
mounted the device.

But sometimes you want to explicitly monitor a NFS share. The
`nfs_space_check` attribute can now be controlled with the puppet type
`nimsoft_disk` with a new `nfs` property